Move `wrap_columns` to its own module
